refactor: 패키지 계층 변경
